taroscript: move still used code over to new file
